restructured package to match RAIL skeleton
