+++ src/rail/yaw_rail/__init__.py
@@ -20,16 +20,3 @@ all negative correlation amplitudes in all generated (spatial) samples to zero.
 This needs refinement in a future release, for now it is advised to use the raw
 clutering redshift estimate from yet_another_wizz.
 """
-
-from .cache import YawCacheCreate, YawCacheDrop
-from .correlation import YawAutoCorrelate, YawCrossCorrelate
-from .summarize import YawSummarize
-from .utils import get_dc2_test_data
-
-__all__ = [
-    "YawAutoCorrelate",
-    "YawCacheCreate",
-    "YawCacheDrop",
-    "YawCrossCorrelate",
-    "YawSummarize",
-]
