+++ src/rail/yaw_rail/utils.py
@@ -11,9 +11,11 @@ from pandas import read_parquet
 
 if TYPE_CHECKING:  # pragma: no cover
     from pandas import DataFrame
+    from rail.core.data import DataHandle
 
 __all__ = [
     "get_dc2_test_data",
+    "handle_has_path",
 ]
 
 
@@ -29,3 +31,8 @@ def get_dc2_test_data() -> DataFrame:
         redshift (`z`).
     """
     return read_parquet("https://portal.nersc.gov/cfs/lsst/PZ/test_dc2_rail_yaw.pqt")
+
+
+def handle_has_path(handle: DataHandle) -> bool:
+    """This is a workaround for a peculiarity of `ceci`."""
+    return handle.path is not None and handle.path != "None"
